Rename P2P validation file to validate
