+++ elements/lisk-p2p/src/p2p.ts
@@ -53,7 +53,7 @@ import {
 	checkPeerCompatibility,
 	outgoingPeerInfoSanitization,
 	sanitizePeerLists,
-} from './validation';
+} from './validate';
 
 import {
 	BASE_10_RADIX,
